requirements.txt
google-generativeai>=0.3.0
google-api-python-client==2.130.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
Flask==3.1.2
flask_cors==6.0.1
pandas==2.3.2
openpyxl==3.1.5